refactor(doctor): extract open-policy allowfrom repair
